Rearrange send command body to avoid stale transactions
Transactions are now built, signed, and sent, as quickly as possible once the user gives the go ahead. This is to avoid stale transactions which used to occur from signing and building a transaction before sending it off while the user double-checks their figures. While that was helpful for seeing what was going on, usually the summary is enough information to determine if this transaction is proper or not anyway.

Add gas command for outputting suggested gas price

Add .gitkeep files to bin folders